refactor(dashboards): move daily sessions trend to engaged community
LFXV2-1468

Daily Sessions Trend belongs with the engagement story, not brand
reach. Moved the chart (and its chart-options) from the Brand Reach
drawer into the Engaged Community drawer, rendered below Community
Breakdown. Engaged Community drawer now takes an optional
brandReachData input carrying the daily trend series.
