+++ apps/lfx-one/src/app/modules/dashboards/executive-director/components/brand-reach-drawer/brand-reach-drawer.component.ts
@@ -5,12 +5,9 @@ import { DecimalPipe } from '@angular/common';
 import { ChangeDetectionStrategy, Component, computed, input, model, Signal } from '@angular/core';
 import { ButtonComponent } from '@components/button/button.component';
 import { CardComponent } from '@components/card/card.component';
-import { ChartComponent } from '@components/chart/chart.component';
-import { lfxColors, MARKETING_SOCIAL_PLATFORM_MAP } from '@lfx-one/shared/constants';
-import { hexToRgba } from '@lfx-one/shared/utils';
+import { MARKETING_SOCIAL_PLATFORM_MAP } from '@lfx-one/shared/constants';
 import { DrawerModule } from 'primeng/drawer';
 
-import type { ChartData, ChartOptions } from 'chart.js';
 import type { BrandReachResponse, BrandReachSocialPlatform } from '@lfx-one/shared/interfaces';
 
 interface SocialPlatformView extends BrandReachSocialPlatform {
@@ -21,7 +18,7 @@ interface SocialPlatformView extends BrandReachSocialPlatform {
 @Component({
   selector: 'lfx-brand-reach-drawer',
   changeDetection: ChangeDetectionStrategy.OnPush,
-  imports: [ButtonComponent, CardComponent, ChartComponent, DecimalPipe, DrawerModule],
+  imports: [ButtonComponent, CardComponent, DecimalPipe, DrawerModule],
   templateUrl: './brand-reach-drawer.component.html',
   styleUrl: './brand-reach-drawer.component.scss',
 })
@@ -52,58 +49,6 @@ export class BrandReachDrawerComponent {
       };
     })
   );
-
-  protected readonly dailyTrendData: Signal<ChartData<'line'>> = computed(() => {
-    const { dailyTrend } = this.data();
-    return {
-      labels: dailyTrend.map((d) => d.day),
-      datasets: [
-        {
-          data: dailyTrend.map((d) => d.sessions),
-          borderColor: lfxColors.blue[500],
-          backgroundColor: hexToRgba(lfxColors.blue[500], 0.1),
-          fill: true,
-          tension: 0.4,
-          borderWidth: 2,
-          pointRadius: 0,
-        },
-      ],
-    };
-  });
-
-  protected readonly dailyTrendOptions: ChartOptions<'line'> = {
-    responsive: true,
-    maintainAspectRatio: false,
-    plugins: {
-      legend: { display: false },
-      tooltip: { enabled: true, mode: 'index', intersect: false },
-    },
-    scales: {
-      x: {
-        display: true,
-        grid: { display: false },
-        ticks: {
-          color: lfxColors.gray[500],
-          font: { size: 11 },
-          maxTicksLimit: 6,
-        },
-      },
-      y: {
-        display: true,
-        grid: { color: lfxColors.gray[200], lineWidth: 1 },
-        border: { display: false },
-        ticks: {
-          color: lfxColors.gray[500],
-          font: { size: 11 },
-          callback: (value) => {
-            const num = Number(value);
-            if (num >= 1_000) return `${(num / 1_000).toFixed(0)}K`;
-            return String(num);
-          },
-        },
-      },
-    },
-  };
 
   protected onClose(): void {
     this.visible.set(false);
