+++ apps/lfx-one/src/app/modules/dashboards/executive-director/components/engaged-community-drawer/engaged-community-drawer.component.html
@@ -155,5 +155,18 @@
         <lfx-chart type="bar" [data]="breakdownChartData()" [options]="breakdownChartOptions" height="100%"></lfx-chart>
       </div>
     </div>
+
+    <!-- Daily Sessions Trend -->
+    @if (brandReachData().dailyTrend.length > 0) {
+      <div class="flex flex-col gap-4 p-4 border border-gray-200 rounded-lg" data-testid="engaged-community-drawer-daily-trend">
+        <div class="flex flex-col gap-1">
+          <h3 class="text-sm font-semibold text-gray-900">Daily Sessions Trend</h3>
+          <p class="text-sm text-gray-600">Website sessions over the last 30 days</p>
+        </div>
+        <div class="h-[240px]" data-testid="engaged-community-drawer-daily-chart">
+          <lfx-chart type="line" [data]="dailyTrendData()" [options]="dailyTrendOptions" height="100%"></lfx-chart>
+        </div>
+      </div>
+    }
   </div>
 </p-drawer>
