+++ apps/lfx-one/src/app/modules/dashboards/executive-director/components/marketing-overview/marketing-overview.component.html
@@ -344,6 +344,7 @@
 <lfx-engaged-community-drawer
   [visible]="activeDrawer() === DashboardDrawerType.NorthStarEngagedCommunity"
   [data]="engagedCommunityData()"
+  [brandReachData]="brandReachData()"
   (visibleChange)="handleDrawerClose()"></lfx-engaged-community-drawer>
 
 <lfx-event-growth-drawer
